LayoutTests/http/tests/contentextensions/sync-xhr-blocked-expected.txt

CONSOLE MESSAGE: line 21: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/sync-xhr-blocked.html from loading a resource from http://127.0.0.1:8000/contentextensions/resources/url-blocking-test.js
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load http://127.0.0.1:8000/contentextensions/resources/url-blocking-test.js. 
Synchronous status: 0, readyState:1, responseText: 
Synchronous status: 0, readyState:4, responseText: 
Synchronous error: Error: NetworkError: DOM Exception 19
